fix not showing mostVoted by default by moving logic to BE
