+++ packages/client/src/components/game/Game.tsx
@@ -52,8 +52,6 @@ interface RoundVotesCount {
 export const Game = () => {
   const [game, setGame] = useState<FullGame | undefined>();
   const [clickedNum, setClickedNum] = useState(null);
-  const [currentRoundVotesCount, setCurrentRoundVotesCount] =
-    useState<RoundVotesCount>({ mostVoted: null, votesCount: new Map() });
   const location = useLocation();
 
   const fiboNums = [0, 1, 2, 3, 5, 8, 13, 21, 34, 55];
@@ -70,15 +68,6 @@ export const Game = () => {
     if (!fullGame.hasOwnProperty('event')) {
       setGame(fullGame);
 
-      // current round revealed: count votes
-      if (fullGame.currentRoundRevealed) {
-        countCurrentRoundVotes(fullGame);
-      }
-
-      if (event === 'reveal') {
-        countCurrentRoundVotes(fullGame);
-      }
-
       if (event === 'complete') {
         setClickedNum(null);
       }
@@ -94,13 +83,11 @@ export const Game = () => {
   const clickNumberEvent = (selectedNumber: number): void => {
     if (game.currentRoundRevealed) {
       // Click number to update final votes
-      if (currentRoundVotesCount.mostVoted) {
+      if (game.currentRoundVotesCount.mostVoted) {
         const newCurrentRoundVotesCount = {
-          ...currentRoundVotesCount,
+          ...game.currentRoundVotesCount,
           mostVoted: selectedNumber,
         };
-
-        setCurrentRoundVotesCount(newCurrentRoundVotesCount);
 
         socket.send({
           event: 'story-event-listener',
@@ -161,36 +148,6 @@ export const Game = () => {
       return user.userRound.hasVoted === false;
     });
 
-  const countCurrentRoundVotes = (event: FullGame): void => {
-    const roundUsers = event.users;
-    const votesCount = new Map<string, number>();
-
-    roundUsers.forEach(user => {
-      const userVote = user.userRound.selectedPoint;
-
-      votesCount[userVote] = votesCount[userVote]
-        ? votesCount[userVote] + 1
-        : 1;
-    });
-
-    // extract key of most voted point from count
-    let max = 0;
-    let mostVoted = '';
-
-    for (const voteKey in votesCount) {
-      if (votesCount[voteKey] >= max) {
-        max = votesCount[voteKey];
-        mostVoted = voteKey;
-      }
-    }
-
-    const roundVotesCount: RoundVotesCount = {
-      mostVoted: parseInt(mostVoted, 10),
-      votesCount: votesCount,
-    };
-
-    setCurrentRoundVotesCount(roundVotesCount);
-  };
   return (
     <div>
       {game ? (
@@ -228,10 +185,8 @@ export const Game = () => {
                   key={'fibo' + num}
                   num={num}
                   index={index}
-                  isMostVoted={
-                    currentRoundVotesCount.mostVoted === num ? true : false
-                  }
-                  voteCount={currentRoundVotesCount.votesCount[num]}
+                  isMostVoted={game.currentRoundVotesCount?.mostVoted === num}
+                  voteCount={game.currentRoundVotesCount?.votesCount[num]}
                   currentRoundRevealed={game.currentRoundRevealed}
                   clickedNum={game.currentRoundRevealed ? null : clickedNum}
                   click={clickNumberEvent}
